cat-recognition-service/pyproject.toml
[project]
name = "gpt-cat-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.23.0",
    "pillow>=10.0.0",
    "torch==2.9.0",
    "torchaudio>=2.0.2",
    "torchvision>=0.15.1",
    "transformers>=4.43.0",
]
